Update complaints.service.ts
Implemented logic of throwing NotFoundExcpetion exception when complaints for amendment and withdraw do not exist.

+++ src/complaints/complaints.service.ts
@@ -1,7 +1,8 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Complaint } from './complaint.entity';
 import { ComplaintsRepository } from './complaints.repository';
+import { ReSubmitComplaintDTO } from './dto/re-submit-complaint.dto';
 import { SubmitComplaintDTO } from './dto/submit-complaint';
 
 @Injectable()
@@ -22,11 +23,19 @@ export class ComplaintsService {
     return this.complaintsRepository.selectCounterComplaints(counteredTo);
   }
 
-  reComplain(submitComplaintDTO: SubmitComplaintDTO): Promise<void> {
-    return this.complaintsRepository.updateComplaint(submitComplaintDTO);
+  async reComplain(reSubmitComplaintDTO: ReSubmitComplaintDTO): Promise<void> {
+    const { idComplaints } = reSubmitComplaintDTO;
+    const complaint = await this.complaintsRepository.findOne(idComplaints);
+    // if complaint doesn't exist
+    if (!complaint)
+      throw new NotFoundException('Subject complaint was not found.');
+    return this.complaintsRepository.updateComplaint(reSubmitComplaintDTO);
   }
 
-  withDrawComplaint(idComplaints: string): Promise<void> {
+  async withDrawComplaint(idComplaints: string): Promise<void> {
+    // if complaint doesn't exist
+    if (!(await this.complaintsRepository.findOne(idComplaints)))
+      throw new NotFoundException('Subject complant was not found');
     return this.complaintsRepository.deleteComplaint(idComplaints);
   }
 }
